+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationEventHandlerServiceTest.java
@@ -1,5 +1,6 @@
 package it.eng.negotiation.service;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
@@ -7,10 +8,13 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Optional;
+import java.util.UUID;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -18,7 +22,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.negotiation.exception.ContractNegotiationAPIException;
+import it.eng.negotiation.exception.ContractNegotiationNotFoundException;
+import it.eng.negotiation.model.Agreement;
 import it.eng.negotiation.model.ContractNegotiation;
+import it.eng.negotiation.model.ContractNegotiationState;
 import it.eng.negotiation.model.ModelUtil;
 import it.eng.negotiation.properties.ContractNegotiationProperties;
 import it.eng.negotiation.repository.AgreementRepository;
@@ -27,6 +34,7 @@ import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
 import it.eng.tools.event.contractnegotiation.ContractNegotiationOfferResponseEvent;
 import it.eng.tools.response.GenericApiResponse;
+import it.eng.tools.util.CredentialUtils;
 
 @ExtendWith(MockitoExtension.class)
 public class ContractNegotiationEventHandlerServiceTest {
@@ -41,17 +49,25 @@ public class ContractNegotiationEventHandlerServiceTest {
 	private OkHttpRestClient okHttpRestClient;
 	@Mock
 	private GenericApiResponse<String> apiResponse;
+	@Mock
+	private CredentialUtils credentialUtils;
 	
 	@InjectMocks
 	private ContractNegotiationEventHandlerService handlerService;
+	
+	@Captor
+	private ArgumentCaptor<ContractNegotiation> argCaptorContractNegotiation;
+	@Captor
+	private ArgumentCaptor<Agreement> argCaptorAgreement;
 	
 	@Test
 	@DisplayName("Handle contract negotiation offer response success")
 	public void handleContractNegotiationOfferResponse_accepted_success() {
 		ContractNegotiationOfferResponseEvent offerResponse = new ContractNegotiationOfferResponseEvent(ModelUtil.CONSUMER_PID, 
 				ModelUtil.PROVIDER_PID, true, Serializer.serializePlainJsonNode(ModelUtil.OFFER));
-		
+		when(properties.getAssignee()).thenReturn(ModelUtil.ASSIGNEE);
 		when(repository.findByProviderPidAndConsumerPid(any(String.class), any(String.class))).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
 		when(apiResponse.isSuccess()).thenReturn(true);
 		// TODO temporary until figure out how to get assignee and assigner
@@ -78,9 +94,10 @@ public class ContractNegotiationEventHandlerServiceTest {
 	@Test
 	@DisplayName("Handle agreement verification message success")
 	public void contractAgreementVerificationMessage_success() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(repository.findByProviderPidAndConsumerPid(any(String.class), any(String.class))).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_AGREED));
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(200);
+		when(apiResponse.isSuccess()).thenReturn(true);
 
 		handlerService.verifyNegotiation(ModelUtil.CONTRACT_AGREEMENT_VERIFICATION_MESSAGE);
 		
@@ -93,7 +110,6 @@ public class ContractNegotiationEventHandlerServiceTest {
 		when(repository.findByProviderPidAndConsumerPid(any(String.class), any(String.class))).thenReturn(Optional.empty());
 
 		assertThrows(ContractNegotiationAPIException.class, () -> handlerService.verifyNegotiation(ModelUtil.CONTRACT_AGREEMENT_VERIFICATION_MESSAGE));
-		
 	}
 	
 	@Test
@@ -102,18 +118,58 @@ public class ContractNegotiationEventHandlerServiceTest {
 		when(repository.findByProviderPidAndConsumerPid(any(String.class), any(String.class))).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_ACCEPTED));
 
 		assertThrows(ContractNegotiationAPIException.class, () -> handlerService.verifyNegotiation(ModelUtil.CONTRACT_AGREEMENT_VERIFICATION_MESSAGE));
-		
 	}
 	
 	@Test
 	@DisplayName("Handle agreement verification message - bad request")
 	public void contractAgreementVerificationMessage_badRequest() {
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
 		when(repository.findByProviderPidAndConsumerPid(any(String.class), any(String.class))).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_AGREED));
 		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
-		when(apiResponse.getHttpStatus()).thenReturn(400);
 		
-		assertThrows(ContractNegotiationAPIException.class, () -> handlerService.verifyNegotiation(ModelUtil.CONTRACT_AGREEMENT_VERIFICATION_MESSAGE));
-		
+		assertThrows(ContractNegotiationAPIException.class, 
+				() -> handlerService.verifyNegotiation(ModelUtil.CONTRACT_AGREEMENT_VERIFICATION_MESSAGE));
 	}
 	
+	@Test
+	@DisplayName("Provider terminate contract negotiation")
+	public void terminateNegotiation() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(repository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(true);
+		
+		handlerService.handleContractNegotiationTerminated(contractNegotaitionId);
+		
+		verify(repository).save(argCaptorContractNegotiation.capture());
+		assertEquals(ContractNegotiationState.TERMINATED, argCaptorContractNegotiation.getValue().getState());
+	}
+	
+	@Test
+	@DisplayName("Provider terminate contract negotiation - contract negotiaton not found")
+	public void terminateNegotiation_cn_not_found() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(repository.findById(contractNegotaitionId)).thenReturn(Optional.empty());
+
+		assertThrows(ContractNegotiationNotFoundException.class,
+				() -> handlerService.handleContractNegotiationTerminated(contractNegotaitionId));
+		
+		verify(repository, times(0)).save(argCaptorContractNegotiation.capture());
+	}
+	
+	@Test
+	@DisplayName("Provider terminate contract negotiation - consumer did not respond")
+	public void terminateNegotiation_consumer_error() {
+		String contractNegotaitionId = UUID.randomUUID().toString(); 
+		when(repository.findById(contractNegotaitionId)).thenReturn(Optional.of(ModelUtil.CONTRACT_NEGOTIATION_REQUESTED));
+		when(credentialUtils.getConnectorCredentials()).thenReturn("credentials");
+		when(okHttpRestClient.sendRequestProtocol(any(String.class), any(JsonNode.class), any(String.class))).thenReturn(apiResponse);
+		when(apiResponse.isSuccess()).thenReturn(false);
+
+		assertThrows(ContractNegotiationAPIException.class,
+				() -> handlerService.handleContractNegotiationTerminated(contractNegotaitionId));
+		
+		verify(repository, times(0)).save(argCaptorContractNegotiation.capture());
+	}
 }